chore: FixedHeadTableScrollerを定義
