+++ packages/smarthr-ui/src/components/Table/client/components/index.ts
@@ -1,0 +1,1 @@
+export { FixedHeadTableScroller } from './FixedHeadTableScroller'
